+++ examples/src/main/scala/example2/CustomerId.scala
@@ -1,0 +1,6 @@
+package example2
+
+import com.sksamuel.avro4s._
+
+object CustomerId { implicit def customerIdRecordFormat = RecordFormat[CustomerId] }
+final case class CustomerId(id: String)
