+++ examples/src/main/scala/example2/ExampleApp.scala
@@ -23,14 +23,12 @@ import com.banno.kafka.admin._
 import com.banno.kafka.schemaregistry._
 import com.banno.kafka.consumer._
 import com.banno.kafka.producer._
-import com.sksamuel.avro4s.RecordFormat
 import org.apache.kafka.clients.admin.NewTopic
 import org.apache.kafka.clients.producer.ProducerRecord
 import scala.concurrent.duration._
 import org.apache.kafka.common.TopicPartition
 
 final class ExampleApp[F[_]: Async: ContextShift] {
-  import ExampleApp._
 
   // Change these for your environment as needed
   val topic = new NewTopic(s"example1.customers.v1", 1, 3.toShort)
@@ -48,8 +46,8 @@ final class ExampleApp[F[_]: Async: ContextShift] {
             address = s"address-${a}",
             priority = a % 3 match {
               case 0 => None
-              case 1 => Some(Gold)
-              case 2 => Some(Platinum)
+              case 1 => Some(Priority.Gold)
+              case 2 => Some(Priority.Platinum)
             }
           )
         )
@@ -110,20 +108,4 @@ final class ExampleApp[F[_]: Async: ContextShift] {
 
 object ExampleApp {
   def apply[F[_]: Async: ContextShift] = new ExampleApp[F]
-}
-
-final case class CustomerId(id: String)
-object CustomerId {
-  implicit val customerIdRecordFormat = RecordFormat[CustomerId]
-}
-sealed trait Priority
-object Priority {
-case object Platinum extends Priority
-case object Gold extends Priority
-
-  implicit val priorityRecordFormat = RecordFormat[Priority]
-}
-final case class Customer(name: String, address: String, priority: Option[Priority] = None)
-object Customer {
-  implicit val customerRecordFormat = RecordFormat[Customer]
 }
